% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getModel.R
\name{CompareModel}
\alias{CompareModel}
\title{Get good model}
\usage{
CompareModel(y, x1, x2)
}
\arguments{
\item{y}{objective variable}

\item{x1}{first explanatory variable}

\item{x2}{second explanatory variable}
}
\value{
good model
}
\description{
Get good model
}

