% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getModel.R
\name{CompareModel}
\alias{CompareModel}
\title{model comparison}
\usage{
CompareModel(y, x1)
}
\arguments{
\item{y}{objective variable}

\item{x1}{first explanatory variable}

}
\description{
model comparison
}

